[PM-33473] Remove `pm-29594-update-individual-subscription-page` feature flag (#7519)
* chore: [PM-33473] remove pm-29594-update-individual-subscription-page feature flag

* chore: dotnet format — add UTF-8 BOM to new response model files
