+++ src/Core/Services/Implementations/UserService.cs
@@ -18,7 +18,6 @@ using Bit.Core.Billing.Licenses.Extensions;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Business;
 using Bit.Core.Billing.Premium.Queries;
-using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Entities;
@@ -68,7 +67,6 @@ public class UserService : UserManager<User>, IUserService
     private readonly ITwoFactorIsEnabledQuery _twoFactorIsEnabledQuery;
     private readonly IDistributedCache _distributedCache;
     private readonly IPolicyRequirementQuery _policyRequirementQuery;
-    private readonly IPricingClient _pricingClient;
     private readonly IHasPremiumAccessQuery _hasPremiumAccessQuery;
     private readonly ISubscriberService _subscriberService;
     private readonly ISendFileStorageService _sendFileStorageService;
@@ -103,7 +101,6 @@ public class UserService : UserManager<User>, IUserService
         ITwoFactorIsEnabledQuery twoFactorIsEnabledQuery,
         IDistributedCache distributedCache,
         IPolicyRequirementQuery policyRequirementQuery,
-        IPricingClient pricingClient,
         IHasPremiumAccessQuery hasPremiumAccessQuery,
         ISubscriberService subscriberService,
         ISendFileStorageService sendFileStorageService)
@@ -142,7 +139,6 @@ public class UserService : UserManager<User>, IUserService
         _twoFactorIsEnabledQuery = twoFactorIsEnabledQuery;
         _distributedCache = distributedCache;
         _policyRequirementQuery = policyRequirementQuery;
-        _pricingClient = pricingClient;
         _hasPremiumAccessQuery = hasPremiumAccessQuery;
         _subscriberService = subscriberService;
         _sendFileStorageService = sendFileStorageService;
@@ -823,33 +819,6 @@ public class UserService : UserManager<User>, IUserService
         await SaveUserAsync(user);
     }
 
-    // TODO: Remove with deletion of pm-29594-update-individual-subscription-page
-    public async Task<string> AdjustStorageAsync(User user, short storageAdjustmentGb)
-    {
-        if (user == null)
-        {
-            throw new ArgumentNullException(nameof(user));
-        }
-
-        if (!user.Premium)
-        {
-            throw new BadRequestException("Not a premium user.");
-        }
-
-        var premiumPlan = await _pricingClient.GetAvailablePremiumPlan();
-
-        var baseStorageGb = (short)premiumPlan.Storage.Provided;
-        var secret = await BillingHelpers.AdjustStorageAsync(
-            _paymentService,
-            null,
-            _featureService,
-            user,
-            storageAdjustmentGb,
-            premiumPlan.Storage.StripePriceId,
-            baseStorageGb);
-        await SaveUserAsync(user);
-        return secret;
-    }
     //TODO: Remove with the deletion of PM32645_DeferPriceMigrationToRenewal feature flag
     public async Task CancelPremiumAsync(User user, bool? endOfPeriod = null)
     {
@@ -860,12 +829,6 @@ public class UserService : UserManager<User>, IUserService
             eop = false;
         }
         await _paymentService.CancelSubscriptionAsync(user, eop);
-    }
-
-    // TODO: Remove with deletion of pm-29594-update-individual-subscription-page
-    public async Task ReinstatePremiumAsync(User user)
-    {
-        await _paymentService.ReinstateSubscriptionAsync(user);
     }
 
     public async Task EnablePremiumAsync(Guid userId, DateTime? expirationDate)
