+++ test/Core.Test/Services/UserServiceTests.cs
@@ -14,7 +14,6 @@ using Bit.Core.Auth.UserFeatures.TwoFactorAuth.Interfaces;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Business;
 using Bit.Core.Billing.Premium.Queries;
-using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Entities;
@@ -573,50 +572,6 @@ public class UserServiceTests
             sutProvider.Sut.AdminResetPasswordAsync(
                 OrganizationUserType.Owner, organization.Id, orgUser.Id, "newPassword", "key"));
         Assert.Equal("Organization User not valid", exception.Message);
-    }
-
-    [Theory, BitAutoData]
-    public async Task AdjustStorageAsync_NullUser_ThrowsArgumentNullException(
-        SutProvider<UserService> sutProvider)
-    {
-        await Assert.ThrowsAsync<ArgumentNullException>(
-            () => sutProvider.Sut.AdjustStorageAsync(null, 1));
-    }
-
-    [Theory, BitAutoData]
-    public async Task AdjustStorageAsync_NotPremium_ThrowsBadRequestException(
-        User user, SutProvider<UserService> sutProvider)
-    {
-        user.Premium = false;
-
-        await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.AdjustStorageAsync(user, 1));
-    }
-
-    [Theory, BitAutoData]
-    public async Task AdjustStorageAsync_Success_CallsPaymentServiceAndSavesUser(
-        User user, SutProvider<UserService> sutProvider)
-    {
-        user.Premium = true;
-        user.GatewayCustomerId = "cus_123";
-        user.GatewaySubscriptionId = "sub_123";
-        user.MaxStorageGb = 1;
-        user.Storage = 0;
-
-        var premiumPlan = new Bit.Core.Billing.Pricing.Premium.Plan
-        {
-            Name = "Premium",
-            Available = true,
-            Seat = new Bit.Core.Billing.Pricing.Premium.Purchasable { StripePriceId = "premium-seat", Price = 10, Provided = 1 },
-            Storage = new Bit.Core.Billing.Pricing.Premium.Purchasable { StripePriceId = "storage-gb-annually", Price = 4, Provided = 1 }
-        };
-
-        sutProvider.GetDependency<IPricingClient>().GetAvailablePremiumPlan().Returns(premiumPlan);
-
-        await sutProvider.Sut.AdjustStorageAsync(user, 1);
-
-        await sutProvider.GetDependency<IStripePaymentService>().Received(1)
-            .AdjustStorageAsync(user, Arg.Any<int>(), premiumPlan.Storage.StripePriceId);
     }
 
     [Theory, BitAutoData]
